+++ src/index.js
@@ -5,9 +5,9 @@ export { default as withDisplayName } from './withDisplayName'
 export { default as withFocusProps } from './withFocusProps'
 export { default as withHoverProps } from './withHoverProps'
 export { default as withJwtProps } from './withJwtProps'
-export { default as withMouseDownProps } from './withMouseDownProps'
 export { default as withNotifyOnLowFPS } from './withNotifyOnLowFPS'
 export { default as withOverrideFromContext } from './withOverrideFromContext'
+export { default as withPressedProps } from './withPressedProps'
 export { default as withStyleSheetOverride } from './withStyleSheetOverride'
 export { default as withTheme } from './withTheme'
 export { default as withTouchProps } from './withTouchProps'
