+++ src/withPressedProps.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import wrapDisplayName from 'recompose/wrapDisplayName'
 
 export default mouseDownProps => Target => {
-  class WithMouseDownProps extends Component {
+  class WithPressedProps extends Component {
     constructor() {
       super()
 
@@ -41,7 +41,7 @@ export default mouseDownProps => Target => {
     }
   }
 
-  WithMouseDownProps.displayName = wrapDisplayName(Target, 'withMouseDownProps')
+  WithPressedProps.displayName = wrapDisplayName(Target, 'withPressedProps')
 
-  return WithMouseDownProps
+  return WithPressedProps
 }
